+++ go/build.sh
@@ -2,9 +2,9 @@
 
 # may have to set GO111MODULE=off
 
-#export GOPATH=$(pwd;)
+# export GOPATH=$(pwd;)
 
-#echo $GOPATH
+# echo $GOPATH
 
 # set -x
 
@@ -14,15 +14,28 @@ install() {
 }
 
 test() {
-    #go test $1 -coverprofile=coverprofile.out  -coverpkg go/src/lib/lomipc,go/src/lib/lomcommon,go/src/vendors/sonic/client/dbclient -covermode=atomic go/src/lib/lib_test,go/src/vendors/sonic/client/dbclient_test
-     go test $1 -coverprofile=coverprofile.out  -coverpkg go/src/vendors/sonic/client/dbclient -covermode=atomic go/src/vendors/sonic/client/dbclient_test
-    if [ $? -ne 0 ]; then
+  go test $1 -coverprofile=coverprofile.out  -coverpkg lib/lomipc,lib/lomcommon -covermode=atomic lib/lib_test
+  if [ $? -ne 0 ]; then
         echo "Failed to run test"
         exit -1
-    fi
-    go tool cover -html=coverprofile.out -o /tmp/coverage.html
-    ls -l coverprofile.out /tmp/coverage.html
-    echo "View /tmp/coverage.html in Edge"
+  fi
+
+  go tool cover -html=coverprofile.out -o /tmp/coverage.html
+  ls -l coverprofile.out /tmp/coverage.html
+  echo "View /tmp/coverage.html in Edge"
+}
+
+etest() { 
+  go test $1 -p 1 -coverprofile=dbclient_coverprofile.out  -coverpkg go/src/vendors/sonic/client/dbclient -covermode=atomic go/src/vendors/sonic/client/dbclient
+  if [ $? -ne 0 ]; then
+  	echo "Failed to run test"
+  	exit -1
+  fi
+  cat ./dbclient_coverprofile.out* > dbclient_coverprofile.out
+  go tool cover -html=dbclient_coverprofile.out -o /tmp/dbclient_coverage.html
+  ls -l dbclient_coverprofile.out /tmp/dbclient_coverage.html
+  echo "View /tmp/dbclient_coverage.html in Edge"
+
 }
 
 clean() {
@@ -53,6 +66,9 @@ elif [[ "list" == "$cmd"* ]]; then
 
 elif [[ "test" == "$cmd"* ]]; then
     test ""
+
+elif [[ "etest" == "$cmd"* ]]; then
+    etest ""
 
 elif [[ "vtest" == "$cmd"* ]]; then
     test "-v"
