+++ go/src/vendors/sonic/client/dbclient/counter_repository.go
@@ -15,25 +15,32 @@ const (
 	if_in_errors_counter_key              string = "IfInErrors"
 	in_unicast_packets_counter_key        string = "InUnicastPackets"
 	out_unicast_packets_counter_key       string = "OutUnicastPackets"
-	base                                  int    = 10
+	atoi_base                                int = 10
+        uint_bit_size                            int = 64
 )
 
 type CounterRepository struct {
 	RedisProvider RedisProviderInterface
 }
 
+/* Cache for storing interface to Oid mapping */
+var interfaceToOidMapping map[string]string
+
 /*
 Returns interface counters for all interfaces on the Sonic device.
-First it gets all oids for interfaces and then gets counters for each interface
+First it gets all oids for interfaces and then gets counters for each interface by performing redis hmGet calls.
 */
 func (counterRepository *CounterRepository) GetInterfaceCounters() (map[string]map[string]uint64, error) {
 
 	var interfaceCountersMap = make(map[string]map[string]uint64)
+        var err error
 
-	interfaceToOidMapping, err := counterRepository.RedisProvider.HGetAll(COUNTER_DB_ID, counters_port_name_map_redis_key)
-	if err != nil {
-		return nil, LogError("HGetAll Failed. err: (%v)", err)
-	}
+	if interfaceToOidMapping == nil {
+	   interfaceToOidMapping, err = counterRepository.RedisProvider.HGetAll(COUNTER_DB_ID, counters_port_name_map_redis_key)
+	   if err != nil {
+       		return nil, LogError("HGetAll Failed. err: (%v)", err)
+           }
+        }
 
 	for interfaceName, interfaceOid := range interfaceToOidMapping {
 
@@ -45,17 +52,17 @@ func (counterRepository *CounterRepository) GetInterfaceCounters() (map[string]m
 			return nil, LogError("HmGet for key (%s) failed. err:(%v)", interfaceCountersKey, err)
 		}
 
-		ifInErrors, err := strconv.ParseUint(result[0].(string), base, 64)
+		ifInErrors, err := strconv.ParseUint(result[0].(string), atoi_base, uint_bit_size)
 		if err != nil {
 			return nil, LogError("IfInErrors counter ParseUint conversion failed for key (%s). err: (%v)", interfaceCountersKey, err)
 		}
 
-		inUnicastPackets, err := strconv.ParseUint(result[1].(string), base, 64)
+		inUnicastPackets, err := strconv.ParseUint(result[1].(string), atoi_base, uint_bit_size)
 		if err != nil {
 			return nil, LogError("InUnicastPackets counter ParseUint conversion failed for key (%s). err: (%v)", interfaceCountersKey, err)
 		}
 
-		outUnicastPackets, err := strconv.ParseUint(result[2].(string), base, 64)
+		outUnicastPackets, err := strconv.ParseUint(result[2].(string), atoi_base, uint_bit_size)
 		if err != nil {
 			return nil, LogError("OutUnicastPackets counter ParseUint conversion failed for key (%s). err: (%v)", interfaceCountersKey, err)
 		}
