+++ go/src/vendors/sonic/client/dbclient/redis_provider.go
@@ -1,7 +1,9 @@
-// This file is reponsible for maintaining connections for redis DBs.
 package dbclient
 
-import "github.com/go-redis/redis"
+import (
+        . "go/src/lib/lomcommon"
+	"github.com/go-redis/redis"
+)
 
 const (
 	redisServerAddr     = "localhost:6379"
@@ -15,13 +17,19 @@ func init() {
 	dbToRedisClientMapping = make(map[int]*redis.Client)
 }
 
-/* Gets redis connection for a specific redis database. The connection is re-used.
-   If multiple go routines call this at same time, the last client will be stored. */
-func GetRedisConnectionForDatabase(databaseIdentifier int) *redis.Client {
+/*
+Gets redis connection for a specific redis database. The connection is re-used.
+If multiple go routines call this at same time, the last client will be stored.
+*/
+func GetRedisConnectionForDatabase(databaseIdentifier int) (*redis.Client, error) {
+	if !validateDbIdentifier(databaseIdentifier) {
+		return nil, LogError("invalid databaseIdentifier (%d)", databaseIdentifier)
+	}
+
 	redisClient, ok := dbToRedisClientMapping[databaseIdentifier]
 
 	if ok {
-		return redisClient
+		return redisClient, nil
 	}
 
 	var client = redis.NewClient(&redis.Options{
@@ -30,8 +38,17 @@ func GetRedisConnectionForDatabase(databaseIdentifier int) *redis.Client {
 		DB:       databaseIdentifier,
 	})
 
+	if client == nil {
+		return nil, LogError("client expected to be non nil for databaseIdentifier (%d)", databaseIdentifier)
+	}
+
 	dbToRedisClientMapping[databaseIdentifier] = client
-	return client
+	return client, nil
+}
+
+/* Validates if databaseIdentifier is allowed to be accessed by LoM" */
+func validateDbIdentifier(databaseIdentifier int) bool {
+	return databaseIdentifier == COUNTER_DB_ID
 }
 
 type RedisProviderInterface interface {
@@ -43,10 +60,28 @@ type RedisProvider struct {
 }
 
 func (redisProvider *RedisProvider) HmGet(database int, key string, fields []string) ([]interface{}, error) {
-	return GetRedisConnectionForDatabase(database).HMGet(key, fields...).Result()
+	client, err := GetRedisConnectionForDatabase(database)
+	if err != nil {
+		return nil, err
+	}
+	return executeHmGet(client, key, fields)
 }
 
 func (redisProvider *RedisProvider) HGetAll(database int, key string) (map[string]string, error) {
-	return GetRedisConnectionForDatabase(database).HGetAll(key).Result()
+	client, err := GetRedisConnectionForDatabase(database)
+	if err != nil {
+		return nil, err
+	}
+	return executeHGetAll(client, key)
 }
 
+func hmGetFunction(redisClient *redis.Client, key string, fields []string) ([]interface{}, error) {
+	return redisClient.HMGet(key, fields...).Result()
+}
+
+func hGetAllFunction(redisClient *redis.Client, key string) (map[string]string, error) {
+	return redisClient.HGetAll(key).Result()
+}
+
+var executeHmGet = hmGetFunction
+var executeHGetAll = hGetAllFunction
